requirements.txt
# requirements.txt
Django==4.2.7
azure-communication-email==1.0.0
dnspython==2.3.0
gunicorn==21.2.0
requests==2.31.0
psycopg2-binary==2.9.6  # If using PostgreSQL